Convert note_offs to long, it is set and compared to other long values that could conceivably exceed the max value of an int.

+++ plugins/pydaw/src/pydaw.h
@@ -140,7 +140,7 @@ typedef struct st_pydaw_data
     //int samples_per_bar;
     float sample_rate;
     long current_sample;  //The sample number of the exact point in the song, 0 == bar0/region0, 44100 == 1 second in at 44.1khz.
-    int note_offs[PYDAW_MAX_TRACK_COUNT][PYDAW_MIDI_NOTE_COUNT];  //When a note_on event is fired, a sample number of when to release it is stored here
+    long note_offs[PYDAW_MAX_TRACK_COUNT][PYDAW_MIDI_NOTE_COUNT];  //When a note_on event is fired, a sample number of when to release it is stored here
     lo_server_thread serverThread;
     
     char * osc_url;
